high_load_systems/Lab3/requirements.txt
neo4j
python-dotenv
